+++ src/generator/client/typescript/pkg/src/decimal_d_ts.rs
@@ -1,0 +1,3 @@
+pub(crate) async fn generate_decimal_d_ts() -> String {
+    include!("decimal.d.ts")
+}