+++ src/generator/client/typescript/pkg/src/decimal_js.rs
@@ -1,0 +1,3 @@
+pub(crate) async fn generate_decimal_js() -> String {
+    include!("decimal.js")
+}